+++ src/main/java/com/hsport/wxprogram/web/controller/planc/FoodController.java
@@ -10,6 +10,7 @@ import io.swagger.annotations.ApiOperation;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
+import javax.servlet.http.HttpServletRequest;
 import java.util.List;
 
 @RestController
@@ -19,7 +20,8 @@ import java.util.List;
 public class FoodController {
     @Autowired
     public IFoodService foodService;
-
+    @Autowired
+    HttpServletRequest request;
     /**
     * 保存和修改公用的
     * @param food  传递的实体
@@ -28,6 +30,10 @@ public class FoodController {
     @ApiOperation(value="新增或修改Food信息")
     @RequestMapping(value="/save",method= RequestMethod.POST)
     public AjaxResult save(@RequestBody Food food){
+        AjaxResult ajaxResult = new AjaxResult();
+        if (!ajaxResult.haveAnyOneLogin(request)){
+            return new AjaxResult("用户无权限或已过期");
+        }
         try {
             if(food.getId()!=null){
                 foodService.updateById(food);
@@ -37,7 +43,7 @@ public class FoodController {
             return AjaxResult.me();
         } catch (Exception e) {
             e.printStackTrace();
-            return AjaxResult.me().setMessage("保存对象失败！"+e.getMessage());
+            return AjaxResult.me().setMessage("保存对象失败！"+e.getMessage()).setSuccess(false);
         }
     }
 
@@ -49,34 +55,20 @@ public class FoodController {
     @ApiOperation(value="删除Food信息", notes="删除对象信息")
     @RequestMapping(value="/{id}",method=RequestMethod.DELETE)
     public AjaxResult delete(@PathVariable("id") Integer id){
+        AjaxResult ajaxResult = new AjaxResult();
+        if (!ajaxResult.haveAnyOneLogin(request)){
+            return new AjaxResult("用户无权限或已过期");
+        }
         try {
             foodService.deleteById(id);
             return AjaxResult.me();
         } catch (Exception e) {
         e.printStackTrace();
-            return AjaxResult.me().setMessage("删除对象失败！"+e.getMessage());
+            return AjaxResult.me().setMessage("删除对象失败！"+e.getMessage()).setSuccess(false);
         }
     }
 
-    //获取用户
-    @ApiOperation(value="根据url的id来获取Food详细信息")
-    @RequestMapping(value = "/{id}",method = RequestMethod.GET)
-    public Food get(@PathVariable("id")Integer id)
-    {
-        return foodService.selectById(id);
-    }
 
-
-    /**
-    * 查看所有的员工信息
-    * @return
-    */
-    @ApiOperation(value="来获取所有Food详细信息")
-    @RequestMapping(value = "/list",method = RequestMethod.GET)
-    public List<Food> list(){
-
-        return foodService.selectList(null);
-    }
 
 
     /**
