+++ src/main/java/com/hsport/wxprogram/web/controller/planc/FoodimgController.java
@@ -12,6 +12,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
+import javax.servlet.http.HttpServletRequest;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.List;
@@ -24,7 +25,8 @@ public class FoodimgController {
 
     @Autowired
     public IFoodimgService foodimgService;
-
+    @Autowired
+    HttpServletRequest request;
     /**
     * 保存和修改公用的
     * @param foodimg  传递的实体
@@ -33,6 +35,10 @@ public class FoodimgController {
     @ApiOperation(value="新增或修改Foodimg信息")
     @RequestMapping(value="/save",method= RequestMethod.POST)
     public AjaxResult save(@RequestBody Foodimg foodimg){
+        AjaxResult ajaxResult = new AjaxResult();
+        if (!ajaxResult.haveAnyOneLogin(request)){
+            return  new AjaxResult("用户无权限或已过期");
+        }
         try {
             if(foodimg.getId()!=null){
                 foodimgService.updateById(foodimg);
@@ -42,7 +48,7 @@ public class FoodimgController {
             return AjaxResult.me();
         } catch (Exception e) {
             e.printStackTrace();
-            return AjaxResult.me().setMessage("保存对象失败！"+e.getMessage());
+            return AjaxResult.me().setMessage("保存对象失败！"+e.getMessage()).setSuccess(false);
         }
     }
 
@@ -54,43 +60,38 @@ public class FoodimgController {
     @ApiOperation(value="删除Foodimg信息", notes="删除对象信息")
     @RequestMapping(value="/{id}",method=RequestMethod.DELETE)
     public AjaxResult delete(@PathVariable("id") Integer id){
+        AjaxResult ajaxResult = new AjaxResult();
+        if (!ajaxResult.haveSysUserLogin(request)){
+            return  new AjaxResult("用户无权限或已过期");
+        }
         try {
             foodimgService.deleteById(id);
             return AjaxResult.me();
         } catch (Exception e) {
         e.printStackTrace();
-            return AjaxResult.me().setMessage("删除对象失败！"+e.getMessage());
+            return AjaxResult.me().setMessage("删除对象失败！"+e.getMessage()).setSuccess(false);
         }
     }
 
     //获取用户
     @ApiOperation(value="根据url的id来获取Foodimg详细信息")
     @RequestMapping(value = "/{id}",method = RequestMethod.GET)
-    public Foodimg get(@PathVariable("id")Integer id)
+    public AjaxResult get(@PathVariable("id")Integer id)
     {
-        return foodimgService.selectById(id);
+        return AjaxResult.me().setResultObj(foodimgService.selectById(id));
     }
-
-
-    /**
-    * 查看所有的员工信息
-    * @return
-    */
-    @ApiOperation(value="来获取所有Foodimg详细信息")
-    @RequestMapping(value = "/list",method = RequestMethod.GET)
-    public List<Foodimg> list(){
-
-        return foodimgService.selectList(null);
-    }
-
 
     @ApiOperation(value = "根据该用户的id查询今天上传的食物图片")
     @RequestMapping(value = "/getFoodListByUserID/{id}", method = RequestMethod.GET)
-    public List<Foodimg> getFoodListByUserID(@PathVariable("id")Integer id) {
+    public AjaxResult getFoodListByUserID(@PathVariable("id")Integer id) {
+        AjaxResult ajaxResult = new AjaxResult();
+        if (!ajaxResult.haveAnyOneLogin(request)){
+            return new AjaxResult("用户无权限或已过期");
+        }
         EntityWrapper<Foodimg> userEntityWrapper = new EntityWrapper<>();
         userEntityWrapper.eq("date", DateUtil.today());
         userEntityWrapper.eq("userID", id);
-        return foodimgService.getFoodListByUserID(id);
+        return AjaxResult.me().setResultObj(foodimgService.getFoodListByUserID(id));
     }
     /**
     * 分页查询数据
@@ -100,35 +101,16 @@ public class FoodimgController {
     */
     @ApiOperation(value="来获取所有Foodimg详细信息并分页", notes="根据page页数和传入的query查询条件 来获取某些Foodimg详细信息")
     @RequestMapping(value = "/json",method = RequestMethod.POST)
-    public PageList<Foodimg> json(@RequestBody FoodimgQuery query)
+    public AjaxResult json(@RequestBody FoodimgQuery query)
     {
+        AjaxResult ajaxResult = new AjaxResult();
+        if (!ajaxResult.haveAnyOneLogin(request)){
+            return new AjaxResult("用户无权限或已过期");
+        }
         Page<Foodimg> page = new Page<Foodimg>(query.getPage(),query.getRows());
             page = foodimgService.selectPage(page);
-            return new PageList<Foodimg>(page.getTotal(),page.getRecords());
+            return AjaxResult.me().setResultObj(new PageList<Foodimg>(page.getTotal(),page.getRecords()));
     }
 
-    @ApiOperation(value="用户上传图片")
-    @RequestMapping(value = "/chuanTu",method = RequestMethod.POST)
-    public AjaxResult chuanTu(@RequestParam("multipartFile") MultipartFile multipartFile){
-        Foodimg foodimg = new Foodimg();
-        String UPLOAD_FOLDER = "D:/images/food";
-        Path path = Paths.get(UPLOAD_FOLDER + "/");
-        //获取当前登录用户  需要修改
-        User user = UserContext.getUser();
-        foodimg.setUserID(user.getId());
-        foodimg.setCoachID(user.getCoachID());
-        foodimg.setDate(DateUtil.today());
-        try {
-            String s = picUtil.singleFileUpload(multipartFile,path);
-            if (s.equals("文件为空，请重新上传")){
-                return AjaxResult.me().setMessage("文件为空，请重新上传！");
-            }
-            foodimg.setFoodImgUrl(s);
-            foodimgService.insert(foodimg);
-            return AjaxResult.me();
-        }catch (Exception e){
-            e.printStackTrace();
-            return AjaxResult.me().setMessage("上传图片失败！"+e.getMessage());
-        }
-    }
+
 }
